python-services/test_engine/requirements.txt
pytest>=7.4.0
requests>=2.31.0
PyYAML>=6.0
allure-pytest>=2.13.0
httpx>=0.25.0
websockets>=12.0
playwright>=1.40.0
locust>=2.20.0
faker>=22.0.0
